✨ feat(data): Folder 관련 DTO 및 모델 데이터 타입 수정 및 필드 변경
- `FolderTreeResponseDTO`의 `folderId`와 `categoryId` 타입을 `Int`에서 `Long`으로 변경
- `FolderSimpleInfo` 모델에서 `tags` 필드를 삭제하고 하위 폴더 구조를 위한 `children` 필드 추가

+++ data/src/main/java/com/linku/data/api/dto/server/FolderTreeResponseDTO.kt
@@ -8,13 +8,13 @@ import com.squareup.moshi.JsonClass
 data class FolderTreeResponseDTO(
 
     @field:Json(name = "folderId")
-    val folderId: Int,
+    val folderId: Long,
 
     @field:Json(name = "folderName")
     val folderName: String,
 
     @field:Json(name = "categoryId")
-    val categoryId: Int,
+    val categoryId: Long,
 
     @field:Json(name = "isBookmarked")
     val isBookmarked: Boolean,
